--- v1/exams.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const bodyParser = require('body-parser');
-const exams = express.Router();
-
-
-exams.get('/', (req, res) => res.status(200).send('Hello World!'));
-
-module.exports = exams;
