--- v1/reviews.js
@@ -1,7 +1,0 @@
-const express = require('express');
-const bodyParser = require('body-parser');
-const reviews = express.Router();
-
-reviews.get('/', (req, res) => res.status(200).send('Hello World!'));
-
-module.exports = reviews;
